% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{fit_pulse}
\alias{fit_pulse}
\title{fit_pulse}
\usage{
fit_pulse(pulse_spec_obj)
}
\arguments{
\item{pulse_spec_obj}{Placeholder}
}
\description{
Primary function for fitting deconvulation model for time series of pulsatile
hormone data
}
\keyword{fit}
\keyword{fit_pulse()}
\keyword{pulse}

