% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsespec.R
\name{print.pulse_spec}
\alias{print.pulse_spec}
\title{Print pulse_spec object}
\usage{
\method{print}{pulse_spec}(x)
}
\arguments{
\item{x}{pulse_spec object}

\item{...}{Other arguments passed on to the individual methods}
}
\description{
This is a generic method that dispatches based on the first argument.
}
\examples{
none currently
print.pulse_spec()
}

