% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{show_args}
\alias{show_args}
\title{show_args}
\usage{
show_args(data)
}
\arguments{
\item{data}{named vector of parameters}
}
\description{
Example via R source code: http://git.io/v0zFx
}
\examples{
none currently
show_args()
}
\keyword{pulse}
\keyword{simulation}

