% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_pulsets}
\alias{simulate_pulsets}
\title{simulate_pulsets}
\usage{
simulate_pulsets(parms_list)
}
\arguments{
\item{parms_list}{Named vector of parameters: T, t, vare, ipimean, ipivar,
ipimin, mua, sda, muw, and sdw are required.  Either muh and varh or
constant.halflife must be specified.  Same for mub and varb or
constant_baseline.}
}
\description{
Simulates hormone concentration per minute for 1 subject.
}
\details{
Pulse simulation function adapted from Ken Horton by reducing to single
  subject, restructuring, and commenting. Ken's originally titled "subject
  model sim aim1 mua.R". You have two options for passing arguments: 1)
  specify arguments individually or 2) pass a named vector of parameters to
  parms_list (see function code for naming scheme).
}
\keyword{pulse}
\keyword{simulate_pulsets()}
\keyword{simulation}

