% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{test_inout}
\alias{test_inout}
\title{test_inout}
\usage{
test_inout(data)
}
\arguments{
\item{data}{Named vector of parameters}
}
\description{
Testing reading in data, converting to C data types and converting back.
  Test List/Dataframe to
}
\examples{
none currently
}
\keyword{pulse}
\keyword{simulation}

