% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{test_inout}
\alias{test_inout}
\title{test_inout}
\usage{
test_inout(x, ...)
}
\arguments{
\item{data}{named vector of parameters}
}
\description{
testing reading in data, converting to c data types and converting back.
  test list/dataframe to
}
\examples{
none currently
test_inout()
}
\keyword{pulse}
\keyword{simulation}

